Меняю int на BigInteger

+++ src/api/pizza_service_api.py
@@ -1,8 +1,6 @@
 from typing import List
 from fastapi import FastAPI, APIRouter, Query
 from pydantic import BaseModel
-from sqlalchemy import BigInteger
-
 from ..service.pizza_service import PizzaService
 from ..model.entities import Pizza, OrderStatus
 from ..model.db import SqlDb
@@ -48,7 +46,7 @@ async def update_order_status(order_id: str, status: int = Query(..., ge=1, le=7
 
 
 @router.post("/user/")
-async def add_user(name: str, phone_number: BigInteger):
+async def add_user(name: str, phone_number: int):
     return pizza_service.add_user(name=name, phone_number=phone_number)
